+++ packages/core/src/components/map-layer/components/map-layer-object.tsx
@@ -1,6 +1,10 @@
 import { Component, ComponentInterface, Event, EventEmitter, Host, Prop, h } from "@stencil/core";
 
 import { DsoSlideToggleCustomEvent } from "../../../components";
+import {
+  Wijzigactie,
+  WrapWijzigactie,
+} from "../../../functional-components/wrap-wijzigactie/wrap-wijzigactie.functional-component";
 import { SlideToggleActiveEvent } from "../../slide-toggle/slide-toggle.interfaces";
 
 import { MapLayerObjectActiveChangeEvent } from "./map-layer-object.interfaces";
@@ -8,6 +12,7 @@ import { MapLayerObjectActiveChangeEvent } from "./map-layer-object.interfaces";
 /**
  * @slot - The label for this Map Layer Object
  * @slot symbol - An optional slot to place a symbol in, representing the Map Layer Object.
+ * @slot label - An optional slot for badges or labels next to the object label.
  */
 @Component({
   tag: "dso-map-layer-object",
@@ -20,6 +25,12 @@ export class MapLayerObject implements ComponentInterface {
    */
   @Prop({ reflect: true })
   active?: boolean;
+
+  /**
+   * An optional 'wijzigactie' that signals if the Map Layer is added or removed.
+   */
+  @Prop({ reflect: true })
+  wijzigactie!: Wijzigactie | undefined;
 
   /**
    * Emitted when user activates or deactivates the Map Layer Object.
@@ -50,21 +61,24 @@ export class MapLayerObject implements ComponentInterface {
   render() {
     return (
       <Host onMouseEnter={() => this.dsoMouseEnter.emit()} onMouseLeave={() => this.dsoMouseLeave.emit()}>
-        <div class="map-layer-object">
-          <div>
-            <slot name="symbol" />
+        <WrapWijzigactie wijzigactie={this.wijzigactie}>
+          <div class="map-layer-object">
+            <div>
+              <slot name="symbol" />
+            </div>
+            <div class="map-layer-object-heading">
+              <slot />
+              <slot name="label" />
+            </div>
+            <div class="slide-toggle-container">
+              <dso-slide-toggle
+                accessibleLabel="Toon op kaart"
+                checked={this.active}
+                onDsoActiveChange={this.handleActiveChange}
+              />
+            </div>
           </div>
-          <div>
-            <slot />
-          </div>
-          <div class="slide-toggle-container">
-            <dso-slide-toggle
-              accessibleLabel="Toon op kaart"
-              checked={this.active}
-              onDsoActiveChange={this.handleActiveChange}
-            />
-          </div>
-        </div>
+        </WrapWijzigactie>
       </Host>
     );
   }